+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.4.1 - 2017-03-01
+- list annotations into the outputchannel instead of the quickpick panel, resolve #13
+- store search result into workspaceState, using the status bar item to show the result at any time
+
 ## 0.4.0 - 2017-02-23
 - list annotations, resolve #7,#9
 - show corresponding message in status bar, resolve #12